Extract UserProfileAttributesList
