+++ portal/src/UserProfileAttributesList.module.scss
@@ -1,0 +1,3 @@
+.fieldName {
+  font-weight: 600;
+}
