+++ portal/src/graphql/portal/StandardAttributesConfigurationScreen.module.scss
@@ -1,7 +1,3 @@
 .widget {
   grid-column: 1 / span 12;
 }
-
-.fieldName {
-  font-weight: 600;
-}
